refactor (pages): scroll logic is getting triggered in dot-pages component
