+++ core-web/apps/dotcms-ui/src/app/portlets/dot-pages/dot-pages-listing-panel/dot-pages-listing-panel.component.spec.ts
@@ -1,5 +1,3 @@
-import { Subject } from 'rxjs';
-
 import { CommonModule } from '@angular/common';
 import { DebugElement } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
@@ -20,19 +18,13 @@ import { of } from 'rxjs/internal/observable/of';
 import { DotAutofocusModule } from '@directives/dot-autofocus/dot-autofocus.module';
 import { DotRelativeDatePipe } from '@dotcms/app/view/pipes/dot-relative-date/dot-relative-date.pipe';
 import { DotMessageService } from '@dotcms/data-access';
-import {
-    CoreWebService,
-    CoreWebServiceMock,
-    DotcmsConfigService,
-    SiteService
-} from '@dotcms/dotcms-js';
+import { CoreWebService, CoreWebServiceMock, DotcmsConfigService } from '@dotcms/dotcms-js';
 import { DotMessagePipe } from '@dotcms/ui';
 import {
     DotcmsConfigServiceMock,
     dotcmsContentletMock,
     dotcmsContentTypeBasicMock,
-    MockDotMessageService,
-    mockSites
+    MockDotMessageService
 } from '@dotcms/utils-testing';
 
 import { DotPagesListingPanelComponent } from './dot-pages-listing-panel.component';
@@ -67,8 +59,6 @@ describe('DotPagesListingPanelComponent', () => {
     let component: DotPagesListingPanelComponent;
     let de: DebugElement;
     let store: DotPageStore;
-
-    const switchSiteSubject = new Subject();
 
     class storeMock {
         get vm$() {
@@ -166,19 +156,7 @@ describe('DotPagesListingPanelComponent', () => {
                     { provide: DotcmsConfigService, useClass: DotcmsConfigServiceMock },
                     { provide: CoreWebService, useClass: CoreWebServiceMock },
                     { provide: DotPageStore, useClass: storeMock },
-                    { provide: DotMessageService, useValue: messageServiceMock },
-                    {
-                        provide: SiteService,
-                        useValue: {
-                            get currentSite() {
-                                return undefined;
-                            },
-
-                            get switchSite$() {
-                                return switchSiteSubject.asObservable();
-                            }
-                        }
-                    }
+                    { provide: DotMessageService, useValue: messageServiceMock }
                 ]
             }).compileComponents();
         });
@@ -196,6 +174,7 @@ describe('DotPagesListingPanelComponent', () => {
             spyOn(store, 'setArchived');
             spyOn(store, 'setSessionStorageFilterParams');
             spyOn(component.goToUrl, 'emit');
+            spyOn(component.pageChange, 'emit');
 
             fixture.detectChanges();
             await fixture.whenStable();
@@ -292,10 +271,11 @@ describe('DotPagesListingPanelComponent', () => {
             );
         });
 
-        it('should reload portlet only when the site change', () => {
-            switchSiteSubject.next(mockSites[0]); // setting the site
-            switchSiteSubject.next(mockSites[1]); // switching the site
-            expect(store.getPages).toHaveBeenCalledWith({ offset: 0 });
+        it('should emit page change', () => {
+            const elem = de.query(By.css('p-table'));
+            elem.triggerEventHandler('onPage');
+
+            expect(component.pageChange.emit).toHaveBeenCalled();
         });
     });
 });
