+++ core-web/apps/dotcms-ui/src/app/portlets/dot-pages/dot-pages-listing-panel/dot-pages-listing-panel.component.ts
@@ -15,10 +15,9 @@ import { LazyLoadEvent } from 'primeng/api';
 import { ContextMenu } from 'primeng/contextmenu';
 import { Table } from 'primeng/table';
 
-import { filter, skip, takeUntil } from 'rxjs/operators';
+import { filter, takeUntil } from 'rxjs/operators';
 
 import { DotMessageService } from '@dotcms/data-access';
-import { SiteService } from '@dotcms/dotcms-js';
 
 import { DotPagesState, DotPageStore } from '../dot-pages-store/dot-pages.store';
 import { DotActionsMenuEventParams } from '../dot-pages.component';
@@ -33,7 +32,7 @@ export class DotPagesListingPanelComponent implements OnInit, OnDestroy, AfterVi
     @ViewChild('table') table: Table;
     @Output() goToUrl = new EventEmitter<string>();
     @Output() showActionsMenu = new EventEmitter<DotActionsMenuEventParams>();
-    @Output() tableScroll = new EventEmitter();
+    @Output() pageChange = new EventEmitter<void>();
 
     private domIdMenuAttached = '';
     private destroy$ = new Subject<boolean>();
@@ -47,11 +46,7 @@ export class DotPagesListingPanelComponent implements OnInit, OnDestroy, AfterVi
         draft: this.dotMessageService.get('Draft')
     };
 
-    constructor(
-        private store: DotPageStore,
-        private dotMessageService: DotMessageService,
-        private dotSiteService: SiteService
-    ) {}
+    constructor(private store: DotPageStore, private dotMessageService: DotMessageService) {}
 
     ngOnInit() {
         this.store.actionMenuDomId$
@@ -73,12 +68,6 @@ export class DotPagesListingPanelComponent implements OnInit, OnDestroy, AfterVi
 
         this.scrollElement?.addEventListener('scroll', () => {
             this.closeContextMenu();
-            this.tableScroll.emit();
-        });
-
-        this.dotSiteService.switchSite$.pipe(takeUntil(this.destroy$), skip(1)).subscribe(() => {
-            this.store.getPages({ offset: 0 });
-            this.scrollToTop(); // To reset the scroll so it shows the data it retrieves
         });
     }
 
@@ -186,12 +175,5 @@ export class DotPagesListingPanelComponent implements OnInit, OnDestroy, AfterVi
         this.store.setArchived(archived);
         this.store.getPages({ offset: 0 });
         this.store.setSessionStorageFilterParams();
-    }
-
-    scrollToTop(): void {
-        this.scrollElement?.scroll({
-            top: 0,
-            left: 0
-        });
     }
 }
