+++ core-web/apps/dotcms-ui/src/app/portlets/dot-pages/dot-pages.component.html
@@ -15,7 +15,7 @@
         <dot-pages-listing-panel
             (goToUrl)="goToUrl($event)"
             (showActionsMenu)="showActionsMenu($event)"
-            (tableScroll)="closeMenu()"
+            (pageChange)="scrollToTop()"
         ></dot-pages-listing-panel>
 
         <dot-add-to-bundle
